Keep enums in one place, group in another file
