+++ output-data/types/GroupedTriggerCodes.ts
@@ -1,0 +1,28 @@
+import TriggerCode, { TriggerCodeGroups } from "./TriggerCode"
+
+const GroupedTriggerCodes: Record<TriggerCodeGroups, TriggerCode[]> = {
+  [TriggerCodeGroups.Bails]: [
+    TriggerCode.TRPR0008,
+    TriggerCode.TRPR0010,
+    TriggerCode.TRPR0020,
+    TriggerCode.TRPR0030
+  ],
+  [TriggerCodeGroups.Custody]: [
+    TriggerCode.TRPR0001,
+    TriggerCode.TRPR0005,
+    TriggerCode.TRPR0006,
+    TriggerCode.TRPR0019,
+    TriggerCode.TRPR0021
+  ],
+  [TriggerCodeGroups.Orders]: [
+    TriggerCode.TRPR0003,
+    TriggerCode.TRPR0016,
+    TriggerCode.TRPR0025,
+    TriggerCode.TRPR0026,
+    TriggerCode.TRPR0029,
+    TriggerCode.TRPS0008
+  ],
+  [TriggerCodeGroups.Warrants]: [TriggerCode.TRPR0002, TriggerCode.TRPR0012]
+}
+
+export default GroupedTriggerCodes
