--- output-data/types/TriggerCodeGroups.ts
@@ -1,8 +1,0 @@
-enum TriggerCodeGroups {
-  Bails = "Bails",
-  Custody = "Custody",
-  Orders = "Orders",
-  Warrants = "Warrants"
-}
-
-export default TriggerCodeGroups
